refactor: 취향리포트 기능 구현

+++ lib/model/common/top_flavor.dart
@@ -7,6 +7,6 @@ part 'top_flavor.freezed.dart';
 class TopFlavor with _$TopFlavor {
   const factory TopFlavor({
     required String flavor,
-    required int percent,
+    required double percent,
   }) = _TopFlavor;
 }
